persistent pins

+++ VirtualTourist/Controller/TravelLocationMapViewController.swift
@@ -16,7 +16,8 @@ class TravelLocationMapViewController: UIViewController, MKMapViewDelegate, UINa
     var photoAlbumViewController: PhotoAlbumViewController!
     var deleteMode = false
     var fetchedResultController: NSFetchedResultsController<Pin>!
-
+    var isTapGesture = false
+    
     @IBOutlet weak var mapView: MKMapView!
     @IBOutlet weak var deleteBarButton: UIBarButtonItem!
     @IBOutlet var longPressGesture: UILongPressGestureRecognizer!
@@ -28,11 +29,13 @@ class TravelLocationMapViewController: UIViewController, MKMapViewDelegate, UINa
         enableTopBarButtons(deleteMode)
         mapView.delegate = self
         photoAlbumViewController = (self.storyboard!.instantiateViewController(withIdentifier: Constants.photoAlbumControllerId) as! PhotoAlbumViewController)
-        setUpFetchedPinsController()
+        fetchedResultController = setupFetchController(createFetchRequest(), delegate: self)
+        loadAnnotations()
     }
     
     override func viewWillAppear(_ animated: Bool) {
-        setUpFetchedPinsController()
+        fetchedResultController = setupFetchController(createFetchRequest(), delegate: self)
+        loadAnnotations()
     }
     
     override func viewDidDisappear(_ animated: Bool) {
@@ -42,20 +45,33 @@ class TravelLocationMapViewController: UIViewController, MKMapViewDelegate, UINa
 
     // MARK: - Map Functions
     func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
-        return getPinViewFromMap(mapView, annotation: annotation, identifier: Constants.pinId, animate: !deleteMode)
+        let map = getPinViewFromMap(mapView, annotation: annotation, identifier: Constants.pinId, animate: isTapGesture)
+        isTapGesture = false
+        return map
     }
     
     func mapView(_ mapView: MKMapView, didSelect view: MKAnnotationView) {
         // remove selection for later "re-selection"
         mapView.deselectAnnotation(view.annotation, animated: false)
-        
+    
         if deleteMode {
-            print("delete mode")
-            mapView.removeAnnotation(view.annotation!)
+            deletePin(annotation: view.annotation!)
         } else {
             photoAlbumViewController.currentLocation = view.annotation!
+            photoAlbumViewController.pin = fetchPinFromMap(annotation: view.annotation!)
             photoAlbumViewController.images = nil
             self.navigationController!.pushViewController(photoAlbumViewController, animated: true)
+        }
+    }
+    
+    private func loadAnnotations() {
+        if let pins = fetchedResultController.fetchedObjects {
+            for pin in pins {
+                let annotation = MKPointAnnotation()
+                annotation.coordinate = CLLocationCoordinate2D(latitude: pin.latitude, longitude: pin.longitude)
+                mapView.removeAnnotation(annotation)
+                mapView.addAnnotation(annotation)
+            }
         }
     }
     
@@ -68,67 +84,32 @@ class TravelLocationMapViewController: UIViewController, MKMapViewDelegate, UINa
     }
     
     func deletePin(annotation: MKAnnotation) {
-        let fetchRequest: NSFetchRequest<Pin> = Pin.fetchRequest()
-        fetchRequest.predicate = NSPredicate(format: "@pin.latitude == @ AND @pin.longitude == @", annotation.coordinate.latitude, annotation.coordinate.longitude)
-        fetchRequest.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: false)]
+        mapView.removeAnnotation(annotation)
         
-        let filterResult = setupFetchController(fetchRequest, cache: Constants.pinId)
-        let pin = filterResult.fetchedObjects![0]
-        
-        let pinToDelete = PersistentContainer.shared.viewContext.object(with: pin.objectID)
-        PersistentContainer.shared.viewContext.delete(pinToDelete)
-        PersistentContainer.shared.saveContext()
-        fetchedResultController = nil
-    }
-    
-    fileprivate func setUpFetchedPinsController() {
-        let fetchRequest: NSFetchRequest<Pin> = Pin.fetchRequest()
-        fetchRequest.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: false)]
-        fetchedResultController = setupFetchController(fetchRequest, cache: Constants.pins)
-    }
-    
-    private func setupFetchController(_ fetchRequest: NSFetchRequest<Pin>, cache: String) -> NSFetchedResultsController<Pin> {
-        fetchedResultController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: PersistentContainer.shared.viewContext, sectionNameKeyPath: nil, cacheName: cache)
-        fetchedResultController.delegate = self
-        do {
-            try fetchedResultController.performFetch()
-        } catch {
-            fatalError("fetch error \(error.localizedDescription)")
+        if let pin = fetchPinFromMap(annotation: annotation) {
+            let pinToDelete = PersistentContainer.shared.viewContext.object(with: pin.objectID)
+            PersistentContainer.shared.viewContext.delete(pinToDelete)
+            PersistentContainer.shared.saveContext()
+        } else {
+            print("pin not found from annotation \(annotation)!!!!!")
         }
-        return fetchedResultController
     }
     
-    // MARK: - Update Core Date
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange sectionInfo: NSFetchedResultsSectionInfo, atSectionIndex sectionIndex: Int, for type: NSFetchedResultsChangeType) {
-           let indexSet = IndexSet(integer: sectionIndex)
-           print("didChange \(indexSet), type = \(type)")
-        /*switch type {
-           case .insert: tableView.insertSections(indexSet, with: .fade)
-           case .delete: tableView.deleteSections(indexSet, with: .fade)
-           default:
-               fatalError("Invalid change type in controller(_:didChange:atSectionIndex:for:). Only .insert or .delete should be possible.")
-           }*/
+    func fetchPinFromMap(annotation: MKAnnotation) -> Pin? {
+        let compoundPredicate = NSCompoundPredicate(type: .and, subpredicates: [
+                NSPredicate(format: "latitude = %@", annotation.coordinate.latitude),
+                NSPredicate(format: "longitude = %@", annotation.coordinate.longitude)])
+            
+        let filteredFetchRequest: NSFetchRequest<Pin> = createFetchRequest(predicate: compoundPredicate)
+        let filtredFetchedResults: NSFetchedResultsController<Pin> = setupFetchController(filteredFetchRequest, delegate: self)
+        
+        return filtredFetchedResults.fetchedObjects?.first
     }
-       
-   func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
-           print("didChange newINde\(newIndexPath), type = \(type), index = \(indexPath)")
-    /*switch type {
-           case .insert:
-               tableView.insertRows(at: [newIndexPath!], with: .fade)
-           case .delete:
-               tableView.deleteRows(at: [indexPath!], with: .fade)
-           case .update:
-               tableView.reloadRows(at: [indexPath!], with: .fade)
-           case .move:
-               tableView.moveRow(at: indexPath!, to: newIndexPath!)
-           default:
-               break
-           }*/
-   }
-    
+
     // MARK: - Delegate Actions
     @IBAction func selectedPointTapped(_ sender: UILongPressGestureRecognizer) {
         if sender.state == .began {
+            isTapGesture = true
             let annotation = convertGestureLocationToMapAnnotation(sender, mapView: mapView)
             addPin(annotation: annotation)
             navigateToLocation(mapView, to: annotation)
